pyproject.toml
[project]
name = "matchatts-implementation-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cython>=3.2.2",
    "diffusers>=0.35.2",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "inflect>=7.5.0",
    "ipywidgets>=8.1.8",
    "librosa>=0.11.0",
    "matcha-tts",
    "matplotlib>=3.10.7",
    "notebook>=7.5.0",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "phonemizer>=3.3.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "rich>=14.2.0",
    "rootutils>=1.0.7",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "unidecode>=1.4.0",
    "wget>=3.2",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
exclude = ["checkpoints*", "synth_output*", "vocoder*", ".venv*"]
